+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Fixed
+
+* Fix `app` type signature in `ASGITransport`. (#3109)
+
 ## 0.27.0 (21st February, 2024)
 
 ### Deprecated